+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Please make sure to add your changes to the appropriate categories:
 - Added `Filter<Timed<T, Tm>>` dt-aware prediction
 - Added `Trapezoidal<T>` dt-aware integrator
 - Added `TimedMean<T>` dt-aware EMA
+- Added `TimedPeakHold<T>` dt-aware decay
 
 ### Changed